# Use `withCallingHandlers()` over `try_fetch()` for performance