Probably my last tidyselect micro-optimization

Motivation is that `try_fetch()` is slower than `withCallingHandlers()` and we aren't using any special `try_fetch()` behavior here:

``` r
bench::mark(
  rlang = try_fetch(1, error = function(cnd) stop("oh no")),
  base = withCallingHandlers(1, error = function(cnd) stop("oh no"))
)
#> # A tibble: 2 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 rlang       14.01µs  15.97µs    60353.        0B     18.1
#> 2 base         1.93µs   2.59µs   382351.        0B     38.2
```

It does change the backtrace _numbers_ a little but that seems ok?

I know `with_subscript_errors()` is always called at least 1 time per `eval_select()` call, and `with_chained_errors()` is called 3 times with the below call, so this ends up adding up to a handful of microseconds:

``` r
library(tidyselect)
library(rlang)

expr <- expr(starts_with("d") | starts_with("m") | starts_with("v"))
bench::mark(eval_select(expr, mtcars), iterations = 50000)

# Dev tidyselect
#> # A tibble: 1 × 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 eval_select(expr, mtcars)    831µs   1.11ms      881.    5.25MB     11.0

# This PR
#> # A tibble: 1 × 6
#>   expression                     min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 eval_select(expr, mtcars)    803µs   1.04ms      951.    5.72MB     11.0
```

Which become a little bit more convincing if you just look at a system time result of 10,000 calls:

```r
library(tidyselect)
library(rlang)
expr <- expr(starts_with("d") | starts_with("m") | starts_with("v"))

# dev tidyselect (run twice)
system.time(for(i in 1:10000) eval_select(expr, mtcars))
#>   user  system elapsed 
#>  9.903   0.055   9.988 
system.time(for(i in 1:10000) eval_select(expr, mtcars))
#>   user  system elapsed 
#>  9.737   0.050   9.820 


# this PR (run twice)
system.time(for(i in 1:10000) eval_select(expr, mtcars))
#>   user  system elapsed 
#>  9.027   0.051   9.106 
system.time(for(i in 1:10000) eval_select(expr, mtcars))
#>    user  system elapsed 
#>  9.255   0.053   9.339 
```